refactor: convert EventLoopBridge subclass into EventLoopPlugin (#307)
Replaces the DefaultMontyBridge subclass EventLoopBridge with a proper
MontyPlugin. Adds a wrapExecuteStream hook to MontyPlugin (opt-in via
hasStreamWrapper) and _streamWrappers plumbing to DefaultMontyBridge so
PluginRegistry can wire stream transformation without requiring subclassing.
Functions renamed el_recv/el_emit to conform to namespace-prefix rules.
