+++ lib/dart_monty_bridge.dart
@@ -7,7 +7,6 @@ library;
 export 'src/bridge/agent_session.dart';
 export 'src/bridge/bridge/bridge_event.dart';
 export 'src/bridge/bridge/bridge_middleware.dart';
-export 'src/bridge/bridge/event_loop_bridge.dart';
 export 'src/bridge/bridge/host_function.dart';
 export 'src/bridge/bridge/host_function_schema.dart';
 export 'src/bridge/bridge/host_param.dart';
@@ -27,6 +26,7 @@ export 'src/bridge/os_call/readonly_fs_provider.dart';
 export 'src/bridge/os_call/sandboxed_fs_provider_stub.dart'
     if (dart.library.io) 'src/bridge/os_call/sandboxed_fs_provider.dart';
 export 'src/bridge/os_call/time_os_provider.dart';
+export 'src/bridge/plugins/event_loop_plugin.dart';
 export 'src/bridge/plugins/message_bus_plugin.dart';
 export 'src/bridge/plugins/sandbox_plugin.dart';
 export 'src/bridge/plugins/template_plugin.dart';
